+++ src/lib/protocols/mail_smtp.c
@@ -73,7 +73,7 @@ void ndpi_search_mail_smtp_tcp(struct ndpi_detection_module_struct *ndpi_struct,
      && (packet->parsed_lines <  NDPI_MAX_PARSE_LINES_PER_PACKET)
      && (ntohs(get_u_int16_t(packet->payload, packet->payload_packet_len - 2)) == 0x0d0a)
      ) {
-    u_int8_t a;
+    u_int16_t a;
     u_int8_t bit_count = 0;
 
     NDPI_PARSE_PACKET_LINE_INFO(ndpi_struct, flow, packet);
